GithubScanner/requirements.txt
PyGithub
bs4
python-dotenv